fixing bugs

+++ code/basic_algorithms/map/reference_calc.h
@@ -3,7 +3,7 @@
 
 void reference_calculation(unsigned int* const h_out,
                            unsigned int* const h_in, 
-                           const size_t NUM_ELEMS) {
+                           const unsigned int NUM_ELEMS) {
   for (unsigned int i = 0; i < NUM_ELEMS; i++) 
     h_out[i] = h_in[i];
 }
